backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy[asyncio]>=2.0.0
asyncpg
alembic
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
httpx